Fixed lack of sign extension on literal for compare instructions.

+++ src/betalib/beta.c
@@ -95,19 +95,19 @@ void beta_tick(Beta* beta) {
 			beta_write_reg(beta, (val_a == val_b)? 1: 0, reg_c);
 			break;
 		case CMPEQC:
-			beta_write_reg(beta, (val_a == literal)? 1: 0, reg_c);
+			beta_write_reg(beta, ((int32_t)val_a == (int32_t)literal)? 1: 0, reg_c);
 			break;
 		case CMPLE:
-			beta_write_reg(beta, (val_a <= val_b)? 1: 0, reg_c);
+			beta_write_reg(beta, ((int32_t)val_a <= (int32_t)val_b)? 1: 0, reg_c);
 			break;
 		case CMPLEC: break;
-			beta_write_reg(beta, (val_a <= literal)? 1: 0, reg_c);
+			beta_write_reg(beta, ((int32_t)val_a <= (int32_t)literal)? 1: 0, reg_c);
 			break;
 		case CMPLT:
-			beta_write_reg(beta, (val_a < val_b)? 1: 0, reg_c);
+			beta_write_reg(beta, ((int32_t)val_a < (int32_t)val_b)? 1: 0, reg_c);
 			break;
 		case CMPLTC:
-			beta_write_reg(beta, (val_a < (int32_t)literal)? 1: 0, reg_c);
+			beta_write_reg(beta, ((int32_t)val_a < (int32_t)literal)? 1: 0, reg_c);
 			break;
 		case DIV:	beta_write_reg(beta, val_a / val_b, reg_c);		break;
 		case DIVC:	beta_write_reg(beta, val_a / literal, reg_c);	break;
